transaction_api/requirements.txt
azure-cosmos==4.14.0
azure-functions==1.24.0
azure-identity==1.25.1
pydantic==2.12.3
python-dotenv==1.2.1
